Add glslify specific testcases
