+++ package.json
@@ -56,6 +56,7 @@
         "eslint-plugin-import": "2.17.2",
         "eslint-plugin-jest": "^22.7.1",
         "eslint-plugin-prettier": "^3.1.0",
+        "glsl-noise": "^0.0.0",
         "husky": "^2.7.0",
         "jasmine-fix": "1.3.1",
         "lint-staged": "^8.2.1",
