+++ spec/fixtures/glslify/import-1.glsl
@@ -1,0 +1,1 @@
+const float b = 2.0;
