+++ spec/fixtures/glslify/import-2.glsl
@@ -1,0 +1,3 @@
+const float d = 4.0;
+#pragma glslify: import(./import-3.glsl)
+const float e = 5.0;
