+++ spec/fixtures/glslify/import-entry.frag
@@ -1,0 +1,9 @@
+precision mediump float;
+const float a = 1.0;
+#pragma glslify: import(./import-1.glsl)
+const float c = 3.0;
+#pragma glslify: import('import-2.glsl')
+
+void main() {
+  gl_FragColor2 = vec4(a, b, c, d);
+}
