+++ spec/fixtures/glslify/include-2.glsl
@@ -1,0 +1,3 @@
+const float d = 4.0;
+#pragma glslify: import("./include-3.glsl")
+const float e = 5.0;
