+++ spec/fixtures/glslify/include-3.glsl
@@ -1,0 +1,3 @@
+const float f = 6.0;
+const float g = 7.0;
+const float h = 8.0;
